Conflicts:
	src/main/java/lu/itrust/business/TS/controller/ControllerAnalysis.java
	src/main/java/lu/itrust/business/TS/controller/ControllerAnalysisStandard.java
	src/main/java/lu/itrust/business/TS/database/dao/hbm/DAOAnalysisStandardHBM.java
	src/main/java/lu/itrust/business/TS/database/dao/hbm/DAOMeasureHBM.java
	src/main/resources/application.properties
	src/main/resources/deployment.properties
	src/main/resources/messages.properties
	src/main/resources/messages_fr.properties
	src/main/webapp/WEB-INF/js/trickservice/actionplan.js
	src/main/webapp/WEB-INF/js/trickservice/reloadSection.js
	src/main/webapp/WEB-INF/views/analyses/single/components/assessment/assets.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/assessment/scenarios.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/estimation/asset.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/estimation/form-cssf.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/estimation/form-normal.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/estimation/scenario.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/soa.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/standards/measure/form.jsp
	src/main/webapp/WEB-INF/views/analyses/single/components/standards/measure/singleMeasure.jsp
	src/main/webapp/WEB-INF/views/analyses/single/menu.jsp

+++ src/main/java/lu/itrust/business/TS/database/service/impl/ServiceWordReportImpl.java
@@ -29,7 +29,7 @@ public class ServiceWordReportImpl implements ServiceWordReport {
 	 * @see lu.itrust.business.TS.database.service.ServiceWordReport#get(java.lang.Integer)
 	 */
 	@Override
-	public WordReport get(Integer id) {
+	public WordReport get(Long id) {
 		return daoWordReport.get(id);
 	}
 
@@ -45,7 +45,7 @@ public class ServiceWordReportImpl implements ServiceWordReport {
 	 * @see lu.itrust.business.TS.database.service.ServiceWordReport#getByIdAndUser(java.lang.Integer, java.lang.String)
 	 */
 	@Override
-	public WordReport getByIdAndUser(Integer id, String username) {
+	public WordReport getByIdAndUser(Long id, String username) {
 		return daoWordReport.getByIdAndUser(id, username);
 	}
 
@@ -105,7 +105,7 @@ public class ServiceWordReportImpl implements ServiceWordReport {
 	 */
 	@Transactional
 	@Override
-	public void delete(Integer id) {
+	public void delete(Long id) {
 		daoWordReport.delete(id);
 	}
